# docs/api/topic-e-plan-2-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinBrain OS Plan 2 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/settings/schema": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings Schema",
        "operationId": "get_settings_schema_settings_schema_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsSchemaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/settings/changes": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Changes",
        "operationId": "list_changes_settings_changes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "applied",
                    "pending_approval",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsChangesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Propose Change",
        "operationId": "propose_change_settings_changes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/changes/{change_id}/approve": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Approve Change",
        "operationId": "approve_change_settings_changes__change_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/changes/{change_id}/reject": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Reject Change",
        "operationId": "reject_change_settings_changes__change_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/rollback": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Rollback",
        "operationId": "rollback_settings_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/settings/templates": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_settings_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/settings/templates/{template_id}/preview": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Preview Template",
        "operationId": "preview_template_settings_templates__template_id__preview_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IndustryTemplateId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/templates/{template_id}/apply": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Apply Template",
        "operationId": "apply_template_settings_templates__template_id__apply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/IndustryTemplateId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsChangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings/message-templates": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_settings_message_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageTemplatesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create Template",
        "operationId": "create_template_settings_message_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/settings/message-templates/{template_id}/approve": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Approve Template",
        "operationId": "approve_template_settings_message_templates__template_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings/alert-rules": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_settings_alert_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRulesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create Rule",
        "operationId": "create_rule_settings_alert_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertRuleRequest": {
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "projected_balance",
              "overdue_amount_per_customer",
              "stock_below_reorder",
              "payroll_coverage_days",
              "marketing_return_per_ringgit",
              "open_disputes"
            ],
            "title": "Metric"
          },
          "operator": {
            "type": "string",
            "enum": [
              "above",
              "below"
            ],
            "title": "Operator"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Threshold"
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/JobFunction"
            },
            "type": "array",
            "maxItems": 11,
            "minItems": 1,
            "title": "Recipients"
          },
          "channel": {
            "type": "string",
            "enum": [
              "in_app",
              "email",
              "telegram"
            ],
            "title": "Channel"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "operator",
          "threshold",
          "recipients",
          "channel"
        ],
        "title": "AlertRuleRequest"
      },
      "AlertRuleResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "rule": {
            "$ref": "#/components/schemas/AlertRule"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "rule"
        ],
        "title": "AlertRuleResponse"
      },
      "AlertRulesResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AlertRule"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "rules"
        ],
        "title": "AlertRulesResponse"
      },
...
      "DataMode": {
        "type": "string",
        "enum": [
          "stub",
          "live"
        ],
        "title": "DataMode",
        "description": "Every contract response says whether its data is a canned stub or computed live."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndustryTemplateId": {
        "type": "string",
        "enum": [
          "trading",
          "services",
          "manufacturing"
        ],
        "title": "IndustryTemplateId"
      },
...
      "JobFunction": {
        "type": "string",
        "enum": [
          "owner",
          "operations",
          "finance",
          "sales",
          "customer_service",
          "marketing",
          "procurement",
          "logistics",
          "production",
          "hr",
          "compliance"
        ],
        "title": "JobFunction",
        "description": "The positions a person can hold. One person may hold several."
      },
...
      "MessageTemplateRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "payment_reminder",
              "customer_reply",
              "supplier_query"
            ],
            "title": "Kind"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "ms",
              "zh"
            ],
            "title": "Language"
          },
          "tone": {
            "type": "string",
            "enum": [
              "formal",
              "friendly"
            ],
            "title": "Tone"
          },
          "body": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "language",
          "tone",
          "body"
        ],
        "title": "MessageTemplateRequest"
      },
      "MessageTemplateResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "template": {
            "$ref": "#/components/schemas/MessageTemplate"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "template"
        ],
        "title": "MessageTemplateResponse"
      },
      "MessageTemplatesResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/MessageTemplate"
            },
            "type": "array",
            "title": "Templates"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "templates"
        ],
        "title": "MessageTemplatesResponse"
      },
...
      "RollbackRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "RollbackRequest"
      },
...
      "SettingsChangeRequest": {
        "properties": {
          "area": {
            "type": "string",
            "enum": [
              "profile",
              "positions",
              "approvals",
              "alerts",
              "financing",
              "security",
              "branding"
            ],
            "title": "Area"
          },
          "value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "area",
          "value"
        ],
        "title": "SettingsChangeRequest"
      },
      "SettingsChangeResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "change": {
            "$ref": "#/components/schemas/SettingsChange"
          },
          "settings": {
            "$ref": "#/components/schemas/TenantSettings"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "change",
          "settings"
        ],
        "title": "SettingsChangeResponse"
      },
      "SettingsChangesResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SettingsChange"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "changes"
        ],
        "title": "SettingsChangesResponse"
      },
      "SettingsResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "template": {
            "$ref": "#/components/schemas/IndustryTemplateId"
          },
          "settings": {
            "$ref": "#/components/schemas/TenantSettings"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "version",
          "template",
          "settings"
        ],
        "title": "SettingsResponse"
      },
      "SettingsSchemaResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "json_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Json Schema"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "json_schema"
        ],
        "title": "SettingsSchemaResponse"
      },
...
      "TemplatePreview": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "template_id": {
            "$ref": "#/components/schemas/IndustryTemplateId"
          },
          "positions_added": {
            "items": {
              "$ref": "#/components/schemas/JobFunction"
            },
            "type": "array",
            "title": "Positions Added"
          },
          "positions_removed": {
            "items": {
              "$ref": "#/components/schemas/JobFunction"
            },
            "type": "array",
            "title": "Positions Removed"
          },
          "designed_positions": {
            "items": {
              "$ref": "#/components/schemas/JobFunction"
            },
            "type": "array",
            "title": "Designed Positions"
          },
          "data_kept": {
            "type": "boolean",
            "title": "Data Kept"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "template_id",
          "positions_added",
          "positions_removed",
          "designed_positions",
          "data_kept"
        ],
        "title": "TemplatePreview"
      },
      "TemplatesResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "current": {
            "$ref": "#/components/schemas/IndustryTemplateId"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/IndustryTemplate"
            },
            "type": "array",
            "title": "Templates"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "current",
          "templates"
        ],
        "title": "TemplatesResponse"
      },
      "TenantSettings": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/CompanyProfile"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionSetting"
            },
            "type": "array",
            "title": "Positions"
          },
          "approvals": {
            "$ref": "#/components/schemas/ApprovalSettings"
          },
          "alerts": {
            "$ref": "#/components/schemas/AlertSettings"
          },
          "financing": {
            "$ref": "#/components/schemas/FinancingPreferences"
          },
          "security": {
            "$ref": "#/components/schemas/SecuritySettings"
          },
          "branding": {
            "$ref": "#/components/schemas/BrandingSettings"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "positions",
          "approvals",
          "alerts",
          "financing",
          "security",
          "branding"
        ],
        "title": "TenantSettings"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}